Major refactor to plugins
* Moved all plugins to plugins folder within src
* Plugins now in charge of handling messages via onText hook
* Added index file to plugins folder to export all onText hooks
* Entry point file now calls all onText hooks exported from index file
* Only run gpt2 on windows machines
* Store all emoji triggers on discord bot
