+++ src/plugins/append.js
@@ -27,7 +27,28 @@ const append = ({fileName, text, message}) => {
   });
 };
 
+const onText = (message) => {
+  const cmd = message.content.split(' ');
+  const botCommand = cmd[0];
+
+  if (botCommand === '!append') {
+    const fileName = cmd[1];
+    let text = cmd.slice(2, cmd.length).join(' ');
+    if (text[0] !== '"' || text[text.length - 1] !== '"') {
+      message.channel.send('Please wrap text in quotation marks.');
+      return;
+    }
+    if (!fileName) {
+      message.channel.send('Please specify a name.');
+      return;
+    }
+    text = text.slice(1, text.length - 1);
+    append({fileName, text, message});
+  }
+};
+
 
 module.exports = {
   append,
+  onText,
 };
