+++ src/plugins/index.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const onTextHooks = [
+  require('./add').onText,
+  require('./post').onText,
+  require('./list').onText,
+  require('./remove').onText,
+  require('./help').onText,
+  require('./markov').onText,
+  require('./rename').onText,
+  require('./play').onText,
+  require('./append').onText,
+  require('./trigger').onText,
+  require('./gpt2').onText,
+];
+
+module.exports = { onTextHooks };
