+++ src/plugins/markov.js
@@ -188,7 +188,43 @@ const postMarkovChannel = (user, channel, origChannel, phrase) => {
   );
 };
 
+const onText = (message) => {
+  const cmd = message.content.split(' ');
+  const botCommand = cmd[0];
+
+  if (botCommand === '!markov') {
+    let string = '';
+    if (cmd[2] !== undefined && cmd[2] !== null) {
+      string = cmd.slice(2, cmd.length).join(' ');
+      if (string[string.length - 1] !== '"') {
+        message.channel.send('Please wrap text in quotation marks.');
+        return;
+      }
+      string = string.slice(1, string.length - 1);
+    }
+    if (string == null) {
+      string = '';
+    }
+    if (cmd[1] === 'all') {
+      markovUser(null, message.guild, message.channel, string);
+      return;
+    }
+    const channel = message.mentions.channels.first();
+    const user = message.mentions.users.first();
+    if (user != null) {
+      markovUser(user, message.guild, message.channel, string);
+      return;
+    }
+    if (channel != null) {
+      markovChannel(channel, message.guild, message.channel, string);
+      return;
+    }
+    message.channel.send('Please specify a User or Channel to markov.');
+  }
+};
+
 module.exports = {
   markovUser,
   markovChannel,
+  onText,
 };
